Update the components to render once only when the page loads
