+++ package.json
@@ -19,6 +19,7 @@
     "react-scripts": "5.0.1",
     "react-test-renderer": "^18.2.0",
     "redux": "^4.2.0",
+    "redux-logger": "^3.0.6",
     "redux-thunk": "^2.4.1",
     "web-vitals": "^2.1.4"
   },
