+++ src/components/Missions.js
@@ -1,17 +1,9 @@
-import React, { useEffect } from 'react';
-import { useDispatch, useSelector } from 'react-redux';
+import { useSelector } from 'react-redux';
 import MissionCard from './Pages/MissionCard';
-import { FetchMissions } from '../redux/missions/MissionSlice';
 import './mission.css';
 
 const Missions = () => {
   const missions = useSelector((state) => state.missions);
-
-  const dispatch = useDispatch();
-
-  useEffect(() => {
-    dispatch(FetchMissions());
-  }, []);
   return (
     <>
       <table className="table">
